fix(page-cache): resolve test failures in PHP 8.5 monorepo environment
- Move helper functions from Pest.php into helpers.php (package-scoped
  require_once) so they load correctly under the root phpunit.xml runner
- Rename createDriver/cleanupDir to createPageCacheFileDriver/cleanupPageCacheDir
  to avoid global function collisions with cache-redis and database tests
- Rename makeRequest/makeResponse to makeCacheCheckerRequest/makeCacheCheckerResponse
  in CacheabilityCheckerTest to avoid collision with layout package tests
- Fix PackageStructureTest dirname depth (2→3) for root composer.json lookup
- Add missing .gitattributes, LICENSE to page-cache and page-cache-file packages
- Add page-cache and page-cache-file to GitHub issue templates
