+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -96,6 +96,8 @@ body:
         - media-imagick
         - notification
         - notification-database
+        - page-cache
+        - page-cache-file
         - pagination
         - pubsub
         - pubsub-pgsql
