+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -84,6 +84,8 @@ body:
         - media-imagick
         - notification
         - notification-database
+        - page-cache
+        - page-cache-file
         - pagination
         - pubsub
         - pubsub-pgsql
